requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-MySQLdb==2.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.2.1
PyMySQL==1.1.0
werkzeug==3.0.1
zipp==3.17.0
